move class to new dirs
