+++ app/src/main/AndroidManifest.xml
@@ -17,16 +17,16 @@
         android:theme="@style/Theme.DefaultBackground"
         tools:targetApi="31">
         <activity
-            android:name=".LibraryActivity"
+            android:name=".ui.library.LibraryActivity"
             android:exported="false" />
         <activity
-            android:name=".SearchActivity"
+            android:name=".ui.tracks.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".ui.settings.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".AudioPlayerActivity"
+            android:name=".ui.audioplayer.AudioPlayerActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
