+++ app/src/main/java/com/example/playlistmaker/MainActivity.kt
@@ -8,6 +8,9 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.playlistmaker.ui.library.LibraryActivity
+import com.example.playlistmaker.ui.settings.SettingsActivity
+import com.example.playlistmaker.ui.tracks.SearchActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
