+++ app/src/main/java/com/example/playlistmaker/SearchHistory.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker
 
 import android.content.SharedPreferences
-import com.example.playlistmaker.network.data.Track
+import com.example.playlistmaker.domain.models.Track
 import com.google.gson.Gson
 
 class SearchHistory(private val sharedPreferences: SharedPreferences) {
