+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackListResponse.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker.data.dto
+
+import com.example.playlistmaker.domain.models.Track
+
+data class TrackListResponse(
+    val resultCount: Int? = null,
+    val results: MutableList<Track>? = null
+)
+
