+++ app/src/main/java/com/example/playlistmaker/ui/settings/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.ui.settings
 
 import android.content.Intent
 import android.net.Uri
@@ -11,7 +11,9 @@ import androidx.appcompat.widget.SwitchCompat
 import androidx.appcompat.widget.Toolbar
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.playlistmaker.App
 import com.example.playlistmaker.App.Companion.DARK_THEME_KEY
+import com.example.playlistmaker.R
 
 class SettingsActivity : AppCompatActivity() {
